+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Coming soon
-- tbd
+- Support for environment variable based configuration
 
 ## [0.10.0] - 2020-11-13
 - Updates compatibility with `io.opentelemetry:opentelemetry-sdk:0.10.0`